Added glow discharge integrated test
