+++ run_tests.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
 
 # Run unit tests
-python3 -m pytest tests/unit_tests -v
+python3 -m pytest -v tests/unit_tests
 
 # Run integrated tests
-# (run individually as FEniCS must reset between runs)
-python3 -m pytest tests/integrated_tests/time_of_flight
-python3 -m pytest tests/integrated_tests/streamer_discharge
+# These must be run independently, as FEniCS must reset between runs.
+# `python3 -m pytest -v tests` will succeed on the first integrated test, but
+# fail on the second/third.
+python3 -m pytest -v tests/integrated_tests/time_of_flight
+python3 -m pytest -v tests/integrated_tests/streamer_discharge
+python3 -m pytest -v tests/integrated_tests/glow_discharge
