+++ tests/integrated_tests/glow_discharge/__init__.py
@@ -1,0 +1,1 @@
+# empty __init__.py
